+++ volmdlr/surfaces.py
@@ -25,7 +25,7 @@ import volmdlr.utils.common_operations as vm_common_operations
 import volmdlr.utils.intersections as vm_utils_intersections
 import volmdlr.utils.parametric as vm_parametric
 from volmdlr import display, edges, grid, wires, curves
-from volmdlr.core import EdgeStyle, point_in_list
+from volmdlr.core import EdgeStyle
 from volmdlr.nurbs.core import evaluate_surface, derivatives_surface, point_inversion, find_multiplicity
 from volmdlr.nurbs.fitting import approximate_surface, interpolate_surface
 from volmdlr.nurbs.operations import split_surface_u, split_surface_v
@@ -1108,8 +1108,7 @@ class Surface3D(DessiaObject):
         for primitive in contour3d.primitives:
             primitive_plane_intersections = self.edge_intersections(primitive)
             for primitive_plane_intersection in primitive_plane_intersections:
-                if not volmdlr.core.point_in_list(primitive_plane_intersection,
-                                                  outer_contour_intersections_with_plane):
+                if not primitive_plane_intersection.in_list(outer_contour_intersections_with_plane):
                     outer_contour_intersections_with_plane.append(primitive_plane_intersection)
         return outer_contour_intersections_with_plane
 
@@ -1197,7 +1196,7 @@ class Surface3D(DessiaObject):
         for curve_plane_intersection in curve_plane_intersections:
             inters = curve_plane_intersection.curve_intersections(curve)
             for intersection in inters:
-                if not volmdlr.core.point_in_list(intersection, intersections):
+                if not intersection.in_list(intersections):
                     intersections.append(intersection)
         return intersections
 
@@ -2548,7 +2547,7 @@ class CylindricalSurface3D(PeriodicalSurface):
                     all_generatrices_intersecting = False
                 for index, point in enumerate(linseg_intersections):
                     if other_surface.point_distance(point) < 1e-6 and \
-                            not volmdlr.core.point_in_list(point, lists_intersections[index]):
+                            not point.in_list(lists_intersections[index]):
                         lists_intersections[index].append(point)
             return lists_intersections, all_generatrices_intersecting
 
@@ -2565,13 +2564,13 @@ class CylindricalSurface3D(PeriodicalSurface):
                 for point in (
                         cylinder_generatrices_point_intersections[0] + cylinder_generatrices_point_intersections[1] +
                         cone_generatrices_point_intersections[0] + cone_generatrices_point_intersections[1]):
-                    if not volmdlr.core.point_in_list(point, intersections_points[0]):
+                    if not point.in_list(intersections_points[0]):
                         intersections_points[0].append(point)
         elif not all_cone_generatrices_intersecting_cylinder:
             intersections_points = [[]]
             for point in (cylinder_generatrices_point_intersections[0] + cylinder_generatrices_point_intersections[1] +
                           cone_generatrices_point_intersections[0] + cone_generatrices_point_intersections[1]):
-                if not volmdlr.core.point_in_list(point, intersections_points[0]):
+                if not point.in_list(intersections_points[0]):
                     intersections_points[0].append(point)
         list_curves = []
         for list_points in intersections_points:
@@ -3383,14 +3382,14 @@ class ToroidalSurface3D(PeriodicalSurface):
         for arc in arcs:
             intersections = cylindrical_surface.circle_intersections(arc)
             for intersection in intersections:
-                if not volmdlr.core.point_in_list(intersection, points_intersections):
+                if not intersection.in_list(points_intersections):
                     points_intersections.append(intersection)
         for edge in cylindrical_surface.get_generatrices(self.outer_radius * 3, 300):
             # \
             #    cylindrical_surface.get_circle_generatrices(72, self.outer_radius * 3):
             intersections = self.edge_intersections(edge)
             for point in intersections:
-                if not volmdlr.core.point_in_list(point, points_intersections):
+                if not point.in_list(points_intersections):
                     points_intersections.append(point)
         return points_intersections
 
@@ -3464,7 +3463,7 @@ class ToroidalSurface3D(PeriodicalSurface):
         for edge in conical_surface.get_generatrices(self.outer_radius * 3, 300):
             intersections = self.edge_intersections(edge)
             for point in intersections:
-                if not volmdlr.core.point_in_list(point, points_intersections):
+                if not point.in_list(points_intersections):
                     points_intersections.append(point)
         return points_intersections
 
@@ -4026,7 +4025,7 @@ class ConicalSurface3D(PeriodicalSurface):
         for gene in cyl_generatrices:
             intersections = spherical_surface.edge_intersections(gene)
             for intersection in intersections:
-                if not volmdlr.core.point_in_list(intersection, intersection_points):
+                if not intersection.in_list(intersection_points):
                     intersection_points.append(intersection)
         return intersection_points
 
@@ -5011,7 +5010,7 @@ class SphericalSurface3D(PeriodicalSurface):
         for gene in cyl_generatrices:
             intersections = spherical_surface.edge_intersections(gene)
             for intersection in intersections:
-                if not volmdlr.core.point_in_list(intersection, intersection_points):
+                if not intersection.in_list(intersection_points):
                     intersection_points.append(intersection)
         return intersection_points
 
@@ -6918,7 +6917,7 @@ class BSplineSurface3D(Surface3D):
         points = []
         for point in linesegment2d.discretization_points(number_points=20):
             point3d = self.point2d_to_3d(point)
-            if not volmdlr.core.point_in_list(point3d, points):
+            if not point3d.in_list(points):
                 points.append(point3d)
         if len(points) < 2:
             return None
@@ -7121,7 +7120,7 @@ class BSplineSurface3D(Surface3D):
         points = []
         for point in bspline_curve2d.discretization_points(number_points=number_points):
             point3d = self.point2d_to_3d(point)
-            if not volmdlr.core.point_in_list(point3d, points):
+            if not point3d.in_list(points):
                 points.append(point3d)
         if len(points) < bspline_curve2d.degree + 1:
             return None
@@ -7137,7 +7136,7 @@ class BSplineSurface3D(Surface3D):
         tol = 1e-6 if arc3d.length() > 1e-5 else 1e-8
         for point3d in arc3d.discretization_points(number_points=number_points):
             point2d = self.point3d_to_2d(point3d, tol)
-            if not volmdlr.core.point_in_list(point2d, points):
+            if not point2d.in_list(points):
                 points.append(point2d)
         start = points[0]
         end = points[-1]
@@ -7283,7 +7282,7 @@ class BSplineSurface3D(Surface3D):
         for point in self.control_points[1:]:
             vector = point - points[0]
             is_colinear = any(vector.is_colinear_to(other_vector) for other_vector in vector_list)
-            if not point_in_list(point, points) and not is_colinear:
+            if not point.in_list(points) and not is_colinear:
                 points.append(point)
                 vector_list.append(vector)
                 if len(points) == 3: